travel-agents/requirements.txt
autogen-agentchat==0.4.2
autogen-ext==0.2.5
openai==1.51.2
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.7.0
requests==2.32.3
httpx==0.27.0
dateparser==1.2.0
rich==13.8.0
